Fixed the memory authstore; Refactor unit tests; Bugfix on delete endpoint
